+++ packages/common-substrate/src/project/versioned/ProjectManifestVersioned.ts
@@ -14,7 +14,7 @@ const SUBSTRATE_SUPPORTED_VERSIONS = {
 
 type Versions = keyof typeof SUBSTRATE_SUPPORTED_VERSIONS;
 
-export type ProjectManifestImpls = InstanceType<typeof SUBSTRATE_SUPPORTED_VERSIONS[Versions]>;
+export type ProjectManifestImpls = InstanceType<(typeof SUBSTRATE_SUPPORTED_VERSIONS)[Versions]>;
 
 export function manifestIsV1_0_0(manifest: ISubstrateProjectManifest): manifest is ProjectManifestV1_0_0Impl {
   return manifest.specVersion === '1.0.0';
@@ -43,7 +43,7 @@ export class SubstrateProjectManifestVersioned implements ISubstrateProjectManif
     return this._impl as ProjectManifestV1_0_0Impl;
   }
 
-  toDeployment(): string | undefined {
+  toDeployment(): string {
     return this._impl.deployment.toYaml();
   }
 
@@ -63,11 +63,11 @@ export class SubstrateProjectManifestVersioned implements ISubstrateProjectManif
     return this._impl.specVersion;
   }
 
-  get description(): string {
+  get description(): string | undefined {
     return this._impl.description;
   }
 
-  get repository(): string {
+  get repository(): string | undefined {
     return this._impl.repository;
   }
 }