+++ packages/types/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Changed
 - Enable strict mode for the other packages.
+## [3.8.0] - 2024-07-01
+### Changed
+- Bump version with `@subql/types-core`
 
 ## [3.7.0] - 2024-06-21
 ### Changed
@@ -240,7 +243,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support block handler
 
-[Unreleased]: https://github.com/subquery/subql/compare/types/3.7.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/types/3.8.0...HEAD
+[3.8.0]: https://github.com/subquery/subql/compare/types/3.7.0...types/3.8.0
 [3.7.0]: https://github.com/subquery/subql/compare/types/3.6.1...types/3.7.0
 [3.6.1]: https://github.com/subquery/subql/compare/types/3.6.0...types/3.6.1
 [3.6.0]: https://github.com/subquery/subql/compare/types/3.5.0...types/3.6.0